+++ frontend/src/app.tsx
@@ -3,9 +3,9 @@ import {AppContextProvider} from "./contexts/app.context";
 import {ProjectContextProvider} from "./contexts/projects.context";
 import {Router} from "./lib/router";
 import {BrowserRouter} from "react-router-dom";
-import {ConfigContextProvider} from "./contexts/connection-types.context.tsx";
+import {ConfigContextProvider} from "./contexts/configContext.context.tsx";
 import {UserContextProvider} from "./contexts/users.context.tsx";
-import {DbBrowserContextProvider} from "./contexts/db-browser.context.tsx";
+import {DbBrowserContextProvider} from "./contexts/dbBrowser.context.tsx";
 
 export function App() {
     return (
