+++ frontend/src/styles/variables.scss
@@ -47,7 +47,7 @@ $spacing-lg: 24px;
 $spacing-xl: 40px;
 
 //borders
-$border-color: $gray-300;
+$border-color: $gray-400;
 $border-radius-small: 4px;
 $border-radius-medium: 8px;
 $border-radius-large: 12px;
